+++ src/objects/jmap.rs
@@ -8,7 +8,7 @@ use crate::{
     },
     signature::{Primitive, ReturnType},
     strings::JNIStr,
-    Env,
+    Env, DEFAULT_LOCAL_FRAME_CAPACITY,
 };
 
 use std::{borrow::Cow, ops::Deref};
@@ -63,8 +63,7 @@ impl JMapAPI {
     ) -> Result<&'static Self> {
         static JMAP_API: OnceCell<JMapAPI> = OnceCell::new();
         JMAP_API.get_or_try_init(|| {
-            let vm = env.get_java_vm();
-            vm.with_env_current_frame(|env| {
+            env.with_local_frame(DEFAULT_LOCAL_FRAME_CAPACITY, |env| {
                 let class = loader_context.load_class_for_type::<JMap>(true, env)?;
                 let class = env.new_global_ref(&class).unwrap();
 
@@ -377,8 +376,7 @@ impl JMapEntryAPI {
     ) -> Result<&'static Self> {
         static JMAPENTRY_API: OnceCell<JMapEntryAPI> = OnceCell::new();
         JMAPENTRY_API.get_or_try_init(|| {
-            let vm = env.get_java_vm();
-            vm.with_env_current_frame(|env| {
+            env.with_local_frame(DEFAULT_LOCAL_FRAME_CAPACITY, |env| {
                 let class = loader_context.load_class_for_type::<JMapEntry>(true, env)?;
                 let class = env.new_global_ref(&class).unwrap();